Standardize position, vel, accel to use float and time

+++ Skystone/Application/Camera.cpp
@@ -5,10 +5,10 @@
 
 #include "Components/Render/SpriteRenderer.h"
 
-Camera::Camera(int width, int height)
+Camera::Camera(float width, float height)
 	:width_(width), height_(height),
-	maxX_(std::numeric_limits<int>::max()),
-	maxY_(std::numeric_limits<int>::max())
+	maxX_(std::numeric_limits<float>::max()),
+	maxY_(std::numeric_limits<float>::max())
 {
 }
 
@@ -16,7 +16,7 @@ Camera::~Camera()
 {
 }
 
-void Camera::setLevelBounds(int maxX, int maxY)
+void Camera::setLevelBounds(float maxX, float maxY)
 {
 	maxX_ = maxX;
 	maxY_ = maxY;
@@ -34,9 +34,9 @@ void Camera::followObject(GameObject& object, float percBehind)
 
 void Camera::stayInLevelBounds()
 {
-	position_.x = std::max(0, position_.x);
+	position_.x = std::max(0.0f, position_.x);
 	position_.x = std::min(position_.x, maxX_ - width_);
 
-	position_.y = std::max(0, position_.y);
+	position_.y = std::max(0.0f, position_.y);
 	position_.y = std::min(position_.y, maxY_ - height_);
 }
